+++ reme-design/reme-dao/src/main/java/com/xumengqi/reme/dao/entity/FavorExample.java
@@ -175,52 +175,52 @@ public class FavorExample {
             return (Criteria) this;
         }
 
-        public Criteria andTargetTypeEqualTo(Byte value) {
+        public Criteria andTargetTypeEqualTo(Integer value) {
             addCriterion("target_type =", value, "targetType");
             return (Criteria) this;
         }
 
-        public Criteria andTargetTypeNotEqualTo(Byte value) {
+        public Criteria andTargetTypeNotEqualTo(Integer value) {
             addCriterion("target_type <>", value, "targetType");
             return (Criteria) this;
         }
 
-        public Criteria andTargetTypeGreaterThan(Byte value) {
+        public Criteria andTargetTypeGreaterThan(Integer value) {
             addCriterion("target_type >", value, "targetType");
             return (Criteria) this;
         }
 
-        public Criteria andTargetTypeGreaterThanOrEqualTo(Byte value) {
+        public Criteria andTargetTypeGreaterThanOrEqualTo(Integer value) {
             addCriterion("target_type >=", value, "targetType");
             return (Criteria) this;
         }
 
-        public Criteria andTargetTypeLessThan(Byte value) {
+        public Criteria andTargetTypeLessThan(Integer value) {
             addCriterion("target_type <", value, "targetType");
             return (Criteria) this;
         }
 
-        public Criteria andTargetTypeLessThanOrEqualTo(Byte value) {
+        public Criteria andTargetTypeLessThanOrEqualTo(Integer value) {
             addCriterion("target_type <=", value, "targetType");
             return (Criteria) this;
         }
 
-        public Criteria andTargetTypeIn(List<Byte> values) {
+        public Criteria andTargetTypeIn(List<Integer> values) {
             addCriterion("target_type in", values, "targetType");
             return (Criteria) this;
         }
 
-        public Criteria andTargetTypeNotIn(List<Byte> values) {
+        public Criteria andTargetTypeNotIn(List<Integer> values) {
             addCriterion("target_type not in", values, "targetType");
             return (Criteria) this;
         }
 
-        public Criteria andTargetTypeBetween(Byte value1, Byte value2) {
+        public Criteria andTargetTypeBetween(Integer value1, Integer value2) {
             addCriterion("target_type between", value1, value2, "targetType");
             return (Criteria) this;
         }
 
-        public Criteria andTargetTypeNotBetween(Byte value1, Byte value2) {
+        public Criteria andTargetTypeNotBetween(Integer value1, Integer value2) {
             addCriterion("target_type not between", value1, value2, "targetType");
             return (Criteria) this;
         }
@@ -355,52 +355,52 @@ public class FavorExample {
             return (Criteria) this;
         }
 
-        public Criteria andStatusEqualTo(Byte value) {
+        public Criteria andStatusEqualTo(Integer value) {
             addCriterion("status =", value, "status");
             return (Criteria) this;
         }
 
-        public Criteria andStatusNotEqualTo(Byte value) {
+        public Criteria andStatusNotEqualTo(Integer value) {
             addCriterion("status <>", value, "status");
             return (Criteria) this;
         }
 
-        public Criteria andStatusGreaterThan(Byte value) {
+        public Criteria andStatusGreaterThan(Integer value) {
             addCriterion("status >", value, "status");
             return (Criteria) this;
         }
 
-        public Criteria andStatusGreaterThanOrEqualTo(Byte value) {
+        public Criteria andStatusGreaterThanOrEqualTo(Integer value) {
             addCriterion("status >=", value, "status");
             return (Criteria) this;
         }
 
-        public Criteria andStatusLessThan(Byte value) {
+        public Criteria andStatusLessThan(Integer value) {
             addCriterion("status <", value, "status");
             return (Criteria) this;
         }
 
-        public Criteria andStatusLessThanOrEqualTo(Byte value) {
+        public Criteria andStatusLessThanOrEqualTo(Integer value) {
             addCriterion("status <=", value, "status");
             return (Criteria) this;
         }
 
-        public Criteria andStatusIn(List<Byte> values) {
+        public Criteria andStatusIn(List<Integer> values) {
             addCriterion("status in", values, "status");
             return (Criteria) this;
         }
 
-        public Criteria andStatusNotIn(List<Byte> values) {
+        public Criteria andStatusNotIn(List<Integer> values) {
             addCriterion("status not in", values, "status");
             return (Criteria) this;
         }
 
-        public Criteria andStatusBetween(Byte value1, Byte value2) {
+        public Criteria andStatusBetween(Integer value1, Integer value2) {
             addCriterion("status between", value1, value2, "status");
             return (Criteria) this;
         }
 
-        public Criteria andStatusNotBetween(Byte value1, Byte value2) {
+        public Criteria andStatusNotBetween(Integer value1, Integer value2) {
             addCriterion("status not between", value1, value2, "status");
             return (Criteria) this;
         }
